Module Mirage_runtime

Mirage run-time utilities.

Release v3.7.5

Log thresholds

type log_threshold = [ `All | `Src of string ] * Logs.level

The type for log threshold.

val set_level : default:Logs.level -> log_threshold list -> unit

set_level ~default l set the log levels needed to have all of the log sources appearing in l be used.

module Arg : sig ... end

include module type of Functoria_runtime with module Mirage_runtime.Arg := Arg

module Key : sig ... end

val with_argv : unit Cmdliner.Term.t list -> string -> string array -> unit

Registering scheduler hooks

val at_exit : (unit -> unit Lwt.t) -> unit

at_exit hook registers hook, which will be executed before the unikernel exits. The first hook registered will be executed last.

val at_enter_iter : (unit -> unit) -> unit

at_enter_iter hook registers hook to be executed at the beginning of each event loop iteration. The first hook registered will be executed last.

If hook calls at_enter_iter recursively, the new hook will run only on the next event loop iteration.

val at_leave_iter : (unit -> unit) -> unit

at_leave_iter hook registers hook to be executed at the end of each event loop iteration. See at_enter_iter for details.

Running hooks

val run_exit_hooks : unit -> unit Lwt.t

run_exit_hooks () calls the sequence of hooks registered with at_exit in sequence.

val run_enter_iter_hooks : unit -> unit

run_enter_iter_hooks () calls the sequence of hooks registered with at_enter_iter in sequence.

val run_leave_iter_hooks : unit -> unit

run_leave_iter_hooks () call the sequence of hooks registered with at_leave_iter in sequence.