Module Mirage_runtime_network

Command-line arguments for network devices

This module is the runtime counter-part of the network command-line arguments defined in Mirage_runtime_arg. Both modules should be kept in sync.

val interface : ?group:string -> ?docs:string -> string -> string Cmdliner.Term.t

A network interface, docs defaults to Mirage_runtime.s_net.

module Arg : sig ... end

Cmdliner.Arg converters for Ipadrr types.

module V4 : sig ... end

Ipv4 Terms

module V6 : sig ... end

Ipv6 Term.ts.

val ipv4_only : ?group:string -> ?docs:string -> unit -> bool Cmdliner.Term.t

An option for dual stack to only use IPv4, docs defaults to Mirage_runtime.s_net.

val ipv6_only : ?group:string -> ?docs:string -> unit -> bool Cmdliner.Term.t

An option for dual stack to only use IPv6, docs defaults to Mirage_runtime.s_net.

val resolver : ?group:string -> ?docs:string -> ?default:string list -> unit -> string list option Cmdliner.Term.t

The address of the DNS resolver to use. See $REFERENCE for format. docs defaults to Mirage_runtime.s_net.

val syslog : ?group:string -> ?docs:string -> Ipaddr.t option -> Ipaddr.t option Cmdliner.Term.t

The address to send syslog frames to, docs defaults to Mirage_runtime.s_log.

val monitor : ?group:string -> ?docs:string -> Ipaddr.t option -> Ipaddr.t option Cmdliner.Term.t

The address to send monitor statistics to, docs defaults to Mirage_runtime.s_log.

val syslog_port : ?group:string -> ?docs:string -> int option -> int option Cmdliner.Term.t

The port to send syslog frames to, docs defaults to Mirage_runtime.s_log.

val syslog_hostname : ?group:string -> ?docs:string -> string -> string Cmdliner.Term.t

The hostname to use in syslog frames, docs defaults to Mirage_runtime.s_log.