Module Mirage_runtime_network.V6

Ipv6 Term.ts.

val network : ?group:string -> ?docs:string -> Ipaddr.V6.Prefix.t option -> Ipaddr.V6.Prefix.t option Cmdliner.Term.t

A network defined by an address and netmask, docs defaults to Mirage_runtime.s_net.

val gateway : ?group:string -> ?docs:string -> Ipaddr.V6.t option -> Ipaddr.V6.t option Cmdliner.Term.t

A default gateway option, docs defaults to Mirage_runtime.s_net.

val accept_router_advertisements : ?group:string -> ?docs:string -> unit -> bool Cmdliner.Term.t

An option whether to accept router advertisements, docs defaults to Mirage_runtime.s_net.