Module Mirage_runtime_network

Command-line arguments for network devices

This module is the runtime counter-part of the network command-line arguments defined in Mirage_runtime_arg. Both modules should be kept in sync.

val interface : ?group:string -> ?docs:string -> string -> string Cmdliner.Term.t

A network interface, docs defaults to Mirage_runtime.s_net.

module Arg : sig ... end

Cmdliner.Arg converters for Ipadrr types.

module V4 : sig ... end

Ipv4 Terms

module V6 : sig ... end

Ipv6 Term.ts.

val ipv4_only : ?group:string -> ?docs:string -> unit -> bool Cmdliner.Term.t

An option for dual stack to only use IPv4, docs defaults to Mirage_runtime.s_net.

val ipv6_only : ?group:string -> ?docs:string -> unit -> bool Cmdliner.Term.t

An option for dual stack to only use IPv6, docs defaults to Mirage_runtime.s_net.

val resolver : ?group:string -> ?docs:string -> ?default:string list -> unit -> string list option Cmdliner.Term.t

The address of the DNS resolver to use. See $REFERENCE for format. docs defaults to Mirage_runtime.s_net.

val dns_servers : ?group:string -> ?docs:string -> string list option -> string list option Cmdliner.Term.t

The address of the DNS servers to use. See $REFERENCE for format. docs defaults to Mirage_runtime.s_net.

val dns_timeout : ?group:string -> ?docs:string -> int64 option -> int64 option Cmdliner.Term.t

The timeout (in nanoseconds) for DNS resolution.

val dns_cache_size : ?group:string -> ?docs:string -> int option -> int option Cmdliner.Term.t

The DNS resolution cache size.

val he_aaaa_timeout : ?group:string -> ?docs:string -> int64 option -> int64 option Cmdliner.Term.t

The timeout (in nanoseconds) for AAAA resolution.

val he_connect_delay : ?group:string -> ?docs:string -> int64 option -> int64 option Cmdliner.Term.t

The delay (in nanoseconds) for establishing connections.

val he_connect_timeout : ?group:string -> ?docs:string -> int64 option -> int64 option Cmdliner.Term.t

The timeout (in nanoseconds) for establishing connections.

val he_resolve_timeout : ?group:string -> ?docs:string -> int64 option -> int64 option Cmdliner.Term.t

The timeout (in nanoseconds) for DNS resolution.

val he_resolve_retries : ?group:string -> ?docs:string -> int option -> int option Cmdliner.Term.t

The number of DNS resolution attempts.

val he_timer_interval : ?group:string -> ?docs:string -> int64 option -> int64 option Cmdliner.Term.t

The interval (in nanoseconds) when the timer is executed.

val ssh_key : ?group:string -> ?docs:string -> string option -> string option Cmdliner.Term.t

The private SSH key.

val ssh_password : ?group:string -> ?docs:string -> string option -> string option Cmdliner.Term.t

The SSH password.

val ssh_authenticator : ?group:string -> ?docs:string -> string option -> string option Cmdliner.Term.t

The SSH authenticator.

val tls_authenticator : ?group:string -> ?docs:string -> string option -> string option Cmdliner.Term.t

The TLS authenticator.

val http_headers : ?group:string -> ?docs:string -> (string * string) list option -> (string * string) list option Cmdliner.Term.t

HTTP headers.

val syslog : ?group:string -> ?docs:string -> Ipaddr.t option -> Ipaddr.t option Cmdliner.Term.t

The address to send syslog frames to, docs defaults to Mirage_runtime.s_log.

val syslog_port : ?group:string -> ?docs:string -> int option -> int Cmdliner.Term.t

The port to send syslog frames to, docs defaults to Mirage_runtime.s_log.

val syslog_truncate : ?group:string -> ?docs:string -> int option -> int option Cmdliner.Term.t

Truncate syslog frames to a specific byte count, docs defaults to Mirage_runtime.s_log.

val syslog_keyname : ?group:string -> ?docs:string -> string option -> string option Cmdliner.Term.t

TLS key used for syslog, docs defaults to Mirage_runtime.s_log.

val monitor : ?group:string -> ?docs:string -> Ipaddr.t option -> Ipaddr.t option Cmdliner.Term.t

The address to send monitor statistics to, docs defaults to Mirage_runtime.s_log.